refactor: extract AccountKeysApi from accounts.rs (#444)
