+++ api/src/openapi.rs
@@ -1,4 +1,5 @@
 pub mod accounts;
+pub mod accounts_keys;
 pub mod admin;
 pub mod allowlist;
 pub mod agents;
@@ -32,6 +33,7 @@ pub mod webhooks;
 pub mod webhooks_disputes;
 
 pub use accounts::AccountsApi;
+pub use accounts_keys::AccountKeysApi;
 pub use admin::AdminApi;
 pub use allowlist::AllowlistApi;
 pub use agents::AgentsApi;
@@ -99,6 +101,7 @@ pub fn create_combined_api() -> impl OpenApi {
             TotpApi,
             RecoveryApi,
             EmailVerificationApi,
+            AccountKeysApi,
         ),
     )
 }
